properly linked the additional controller
